Rename auth folder to lowercase
Auth isn't a component by itself
